Optimisation
Change the way we check if SecureRetract State is On. Now we check it
directly in initSecureRectract()
That lighter the code, we don't check it in each functions any more.
We still force The state of SecureRetract for escaping fight (if it
disable, we enable it and switch back to Off) ; all we want is to escape
fight so..