Add ZoomScale() and JoystickAngle Settings
When simulate a rotation of the handle for a center mounted stick. If
you use high angles, you may notice that it becomes impossible to reach
the corner of the axes range of values. So, we add Angle setting and let
the script calculate the right scale value with ZoomScale() .