Add Full Explorer Mode System Jump (Not Yet Fully Tested so take care :) )
Put a WARNING in Readme to be sure to use ENG(UK) Keyboard before making change in the keybindings In Game. I'm not from UK but I always use UK Keyboard so I don't know if all this script work if you use any other Keyboard layout in your operating system. BUT I'm sure ED Fuck' the bindings file if you are in another keyboard layout and make change In Game...

Add AutoDiscoveryScan & TargetNearestStar Options and notes for Explorer System Jump in Settings File
Also added initExplorerSystemScan() & initTargetNearestStar() Functions (in CORE File)
Add Target Nearest Star Macro (you need to be in the navigation panel to make it work)
Add the description of the change and how it work in the Readme in EXPLORER Section