Start To Trying to Get ED states From status.json
I do not feel comfortable with all of this, but if I get it working, it will be a big plus for tracking states of stuff from the game. That should simplify the script a lot.

I need Help to parse and send data from status.json  information overs TCP or UDP, If someone ear me ....

Adding Socket Stuff
Adding Flags bit from status.json